+++ cudpp/src/app/compact_app.cu
@@ -105,7 +105,7 @@ void calculatCompactLaunchParams(const unsigned int numElements,
   */
 template<class T>
 void compactArray(T                      *d_out, 
-                  size_t                 *d_numValidElements,
+                  unsigned int           *d_numValidElements,
                   const T                *d_in, 
                   const unsigned int     *d_isValid,
                   size_t                 numElements,
@@ -190,7 +190,7 @@ void freeCompactStorage(CUDPPCompactPlan *plan)
  * 
  */
 void cudppCompactDispatch(void                   *d_out, 
-                          size_t                 *d_numValidElements,
+                          unsigned int           *d_numValidElements,
                           const void             *d_in, 
                           const unsigned int     *d_isValid,
                           size_t                 numElements,
