+++ cudpp/src/cudpp.cpp
@@ -255,7 +255,7 @@ CUDPPResult cudppMultiScan(CUDPPHandle         planHandle,
 CUDPP_DLL
 CUDPPResult cudppCompact(CUDPPHandle        planHandle,
                          void               *d_out, 
-                         size_t             *d_numValidElements,
+                         unsigned int       *d_numValidElements,
                          const void         *d_in, 
                          const unsigned int *d_isValid,
                          size_t             numElements,
